backend/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
python-jose
python-dotenv
requests
groq
pydantic[email]